requirements.txt
Django==1.7.3
django-extensions==1.4.9
django-hstore==1.3.5
django-jsonfield==0.9.13
psycopg2==2.5.4
